# web/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_create_task_api_tasks_post": {
        "properties": {
          "area_file": {
            "contentMediaType": "application/octet-stream",
            "title": "Area File",
            "type": "string"
          },
          "cloud_cover_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cloud Cover Pct"
          },
          "criterion_alpha": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criterion Alpha"
          },
          "criterion_mode": {
            "title": "Criterion Mode",
            "type": "string"
          },
          "environment_id": {
            "title": "Environment Id",
            "type": "string"
          },
          "gsd_cm": {
            "title": "Gsd Cm",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "survey_type": {
            "title": "Survey Type",
            "type": "string"
          },
          "wind_speed_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind Speed Ms"
          },
          "window_end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window End"
          },
          "window_start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window Start"
          },
          "work_date": {
            "format": "date",
            "title": "Work Date",
            "type": "string"
          }
        },
        "required": [
          "name",
          "environment_id",
          "survey_type",
          "gsd_cm",
          "work_date",
          "criterion_mode",
          "area_file"
        ],
        "title": "Body_create_task_api_tasks_post",
        "type": "object"
      },
...
      "Body_update_task_api_tasks__task_id__put": {
        "properties": {
          "area_file": {
            "contentMediaType": "application/octet-stream",
            "title": "Area File",
            "type": "string"
          },
          "cloud_cover_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cloud Cover Pct"
          },
          "criterion_alpha": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criterion Alpha"
          },
          "criterion_mode": {
            "title": "Criterion Mode",
            "type": "string"
          },
          "expected_version": {
            "title": "Expected Version",
            "type": "integer"
          },
          "gsd_cm": {
            "title": "Gsd Cm",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "survey_type": {
            "title": "Survey Type",
            "type": "string"
          },
          "wind_speed_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind Speed Ms"
          },
          "window_end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window End"
          },
          "window_start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window Start"
          },
          "work_date": {
            "format": "date",
            "title": "Work Date",
            "type": "string"
          }
        },
        "required": [
          "expected_version",
          "name",
          "survey_type",
          "gsd_cm",
          "work_date",
          "criterion_mode",
          "area_file"
        ],
        "title": "Body_update_task_api_tasks__task_id__put",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TaskDetail": {
        "properties": {
          "area": {
            "additionalProperties": true,
            "title": "Area",
            "type": "object"
          },
          "cloud_cover_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cloud Cover Pct"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "criterion_alpha": {
            "title": "Criterion Alpha",
            "type": "number"
          },
          "criterion_mode": {
            "enum": [
              "Время",
              "Налет",
              "Компромисс"
            ],
            "title": "Criterion Mode",
            "type": "string"
          },
          "daylight_warning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daylight Warning"
          },
          "environment_id": {
            "title": "Environment Id",
            "type": "string"
          },
          "environment_name": {
            "title": "Environment Name",
            "type": "string"
          },
          "gsd_cm": {
            "title": "Gsd Cm",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "enum": [
              "Черновик",
              "Рассчитана",
              "Подтверждена"
            ],
            "title": "Status",
            "type": "string"
          },
          "survey_type": {
            "enum": [
              "RGB",
              "мультиспектральная",
              "ИК",
              "LiDAR",
              "геофизическая"
            ],
            "title": "Survey Type",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "wind_speed_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind Speed Ms"
          },
          "window_end": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window End"
          },
          "window_start": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window Start"
          },
          "work_date": {
            "format": "date",
            "title": "Work Date",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "environment_id",
          "environment_name",
          "survey_type",
          "work_date",
          "status",
          "version",
          "created_at",
          "updated_at",
          "gsd_cm",
          "window_start",
          "window_end",
          "wind_speed_ms",
          "cloud_cover_pct",
          "criterion_mode",
          "criterion_alpha",
          "area"
        ],
        "title": "TaskDetail",
        "type": "object"
      },
      "TaskSummary": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "daylight_warning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daylight Warning"
          },
          "environment_id": {
            "title": "Environment Id",
            "type": "string"
          },
          "environment_name": {
            "title": "Environment Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "enum": [
              "Черновик",
              "Рассчитана",
              "Подтверждена"
            ],
            "title": "Status",
            "type": "string"
          },
          "survey_type": {
            "enum": [
              "RGB",
              "мультиспектральная",
              "ИК",
              "LiDAR",
              "геофизическая"
            ],
            "title": "Survey Type",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "work_date": {
            "format": "date",
            "title": "Work Date",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "environment_id",
          "environment_name",
          "survey_type",
          "work_date",
          "status",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "TaskSummary",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Галс — планировщик БВС",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/tasks": {
      "get": {
        "operationId": "list_tasks_api_tasks_get",
        "parameters": [
          {
            "in": "query",
            "name": "environment_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Environment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskSummary"
                  },
                  "title": "Response List Tasks Api Tasks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Tasks",
        "tags": [
          "tasks"
        ]
      },
      "post": {
        "operationId": "create_task_api_tasks_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Name"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_task_api_tasks_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Task",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/tasks/{task_id}": {
      "get": {
        "operationId": "get_task_api_tasks__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Task",
        "tags": [
          "tasks"
        ]
      },
      "put": {
        "operationId": "update_task_api_tasks__task_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Name"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_update_task_api_tasks__task_id__put"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Task",
        "tags": [
          "tasks"
        ]
      }
    }
  }
}